Revert "Fixed printing addresses using a long instead of an int"
This reverts commit 00ee414ff6b57e5df0ee3da612ef8f5dcf0d6019.

+++ print_base.c
@@ -6,7 +6,7 @@
 * @upper: Selects uppercase mode for hex.
 * Return: Number of characters printed.
 */
-int print_base(unsigned long number, short base, short upper)
+int print_base(unsigned int number, short base, short upper)
 {
 	char ans[32], *comp;
 	char low[] = {'0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'b', 'c',
